fix: [#445] use markPrice if there is no candle last
